requirements.txt
gradio>=4.44.1
requests>=2.31.0
Pillow>=10.1.0
duckduckgo-search>=8.0.1
pydantic==2.10.6
fastapi>=0.95.2
starlette>=0.27.0
ollama>=0.4.8